+++ CHANGELOG.md
@@ -5,7 +5,13 @@ Changelog
 
 # 1.x release
 
-## v1.4.1 (master)
+## v1.5.0 (master)
+
+- Enhance: rejected promise now use custom `Error` (thx to @pekeler)
+- Enhance: `FetchError` contains `err.type` and `err.code`, allows for better error handling (thx to @pekeler)
+- Enhance: basic support for redirect mode `manual` and `error`, allows for location header extraction (thx to @jimmywarting for the initial PR)
+
+## v1.4.1
 
 - Fix: wrapping Request instance with FormData body again should preserve the body as-is